src/requirements.txt
pyftpdlib==2.0.1
python-telegram-bot==22.3
requests==2.25.1